Remove all references to survey results
This is a first step towards splitting the code into the common functionality,
which deals with sensed and analysed results, and the study-specific
functionality, which will go into another module in the future.

This removes the existing implementations of `query_trip_ends`,
`query_user_profile_survey`, and `tidy_trip_ends`, and all references to them.

The one place where we didn't just remove the reference was in
`most_recent_trip_ends`.  The subsequent code deals with missing recent trip
ends, so we set `most_recent_trip_ends` to an empty vector instead of removing
the subsequent references.