Drop only columns that exist
The test database doesn't have the `device_token` and `curr_sync_interval` entries in the profile.
Maybe because I loaded the user through the command line and used a test client
that didn't have push configured correctly.

Let's generalize this by only purging the columns that need to be purged. Maybe
other users will choose to not integrate with push notifications in the future.

Thanks to @ahcyip for the tip about `eval` to use the vector variable instead of constant