Remove the conversion to datetime to prevent equisse from automatically filtering
Without this change, the `update_ts` column was a `datetime` object instead of
a string. It looks like this causes equisse to automatically filter based on it.

Without this change, the Export & Code for the participant table was

```
data_esquisse %>%
 filter(update_ts >= "2020-10-14 16:05:42" & update_ts <= "2020-10-14 17:09:59") %>%
```
which causes the last participant entry to be skipped.

With this change, the Export & Code for the participant table is blank, similar
to the value for the Trips tab.

So nothing is filtered, the data popup shows 2/2 entries, and both entries show up in the visualization.