Remove all references to survey results (#4) (#8)
* Remove all references to survey results (#4)

* Remove all references to survey results

This is a first step towards splitting the code into the common functionality,
which deals with sensed and analysed results, and the study-specific
functionality, which will go into another module in the future.

This removes the existing implementations of `query_trip_ends`,
`query_user_profile_survey`, and `tidy_trip_ends`, and all references to them.

The one place where we didn't just remove the reference was in
`most_recent_trip_ends`.  The subsequent code deals with missing recent trip
ends, so we set `most_recent_trip_ends` to an empty vector instead of removing
the subsequent references.

* ci: replace tic's GHA templates with a usethis's GHA template. #5 

Fix #3

* try a GHA template from the usethis package

* run the test ci template on this test branch

* add a GHA CI badge to README

* remove tic's GHA templates

* remove the CI trigger of this test branch

* Add a sample database for dev and test purposes

In response to https://github.com/asiripanich/emdash/issues/2#issuecomment-708047628
It contains background sensed data, analysed data and user annotations consistent with master.
More details in the README

* add a shell script to quickly start the test database container

* add an instruction to use the test database on to README

* add data/dump to .gitignore

* generalise functions in mod_load_data.R to work with the test database

* add unit tests

* Drop the "geometry" column correctly

Instead of using `-"geometry"`, which doesn't work according to
https://r.789695.n4.nabble.com/quot-invalid-argument-to-unary-operator-quot-while-selecting-rows-by-name-td4724899.html

use `dplyr:select` as outlined in
https://www.marsja.se/how-to-remove-a-column-in-r-using-dplyr-by-name-and-index/

* Drop only columns that exist

The test database doesn't have the `device_token` and `curr_sync_interval` entries in the profile.
Maybe because I loaded the user through the command line and used a test client
that didn't have push configured correctly.

Let's generalize this by only purging the columns that need to be purged. Maybe
other users will choose to not integrate with push notifications in the future.

Thanks to @ahcyip for the tip about `eval` to use the vector variable instead of constant

* Remove the `st_drop_geometry` line

Since it is called on trips, which doesn't appear to be an st object.

* Return the cleaned trips table and revert dropping the geometry

Return the cleaned_trips_table from `tidy_cleaned_trips`
Otherwise everything broke downstream

Also revert 1c28ef78138d0db5e6c209d5d20584ad965c6f79
since it was also caused by the fact that the trips object was NULL
the fact that it was not an st object is just icing on the cake :)

* Comment out the branch box from the UI

Since it doesn't actually work unless you are doing an RCT

* Remove the conversion to datetime to prevent equisse from automatically filtering

Without this change, the `update_ts` column was a `datetime` object instead of
a string. It looks like this causes equisse to automatically filter based on it.

Without this change, the Export & Code for the participant table was

```
data_esquisse %>%
 filter(update_ts >= "2020-10-14 16:05:42" & update_ts <= "2020-10-14 17:09:59") %>%
```
which causes the last participant entry to be skipped.

With this change, the Export & Code for the participant table is blank, similar
to the value for the Trips tab.

So nothing is filtered, the data popup shows 2/2 entries, and both entries show up in the visualization.

* Add data for a second user 

This is data from 2015
Mark this user as using `ios` to have some variation in the outputs
Updated the README

Cherry-picked from c8d5290ef39debf621b8cb077607d686c0a75123

* move test data to test-data folder 

data is a reserved folder name in R packages. This triggers a warning when building the package when running R CMD check.

Co-authored-by: amarin <17020181+asiripanich@users.noreply.github.com>

* remove a non imported package from a test script

Co-authored-by: shankari <shankari@eecs.berkeley.edu>