Add an option to anonymize the location information
Our current users have a request to remove the linkage between the trajectory
information in the map view and the uuid.

They don't collect home information in their pre or post surveys, so removing
the link between the UUID and the trajectories will ensure that they cannot
reidentify participants through their the home and work locations. They can
potentially determine the home and work locations of people who have installed
the app, but they cannot link this information with the survey data.

This may or may not be a frequent use case, so I have exposed this via a startup option.

Use `:run_app(mongo_url = 'mongodb://host.docker.internal:27017',
anon_locations=TRUE)` if you wish to anonymize the locations.

Testing done:
- Ran without `anon_locations` = TRUE, mapview had the UUIDs
- Ran with `anon_locations` = TRUE, mapview had `user_1`, `user_2`,... instead