Fix the assignment of the anonymous userids to the trip data table
https://github.com/asiripanich/emdash/pull/11/commits/d2bd57bdc6b9e9fe210cf8a359602043cea3b95c
won't work because the trips table may have 200 entries from only 5 users.
We need to map each uuid to the corresponding anon uuid.

If you don't do that, you get the error

```
Warning: Error in [[<-.data.frame: replacement has 5 rows, data has 12
```

like I just did 😄

If you can think of a better way of doing this that is not a for loop, that would be great, but let's get it working for now.