Display user labels generically in the dashboard
After doing the heavy lifting for implementing
https://github.com/e-mission/e-mission-docs/issues/476
this is actually super simple

- Switch to reading confirmed_trips instead of cleaned_trips
- This leads to a table with user_inputXXX for each user input XXX
    - Remove the user input prefix

We now end up with a table with the user inputs as columns.
Et Voila!

Please see screenshots in PR

This doesn't fully solve the problem of how to determine whether trips have
been confirmed or not, but that will require some careful design to be sufficiently general.
I will open an issue for that and send a separate PR.