Display the inferred labels correctly
This fixes https://github.com/asiripanich/emdash/issues/69

It basically changes the rendering for the inferred trips table to JSON stringify the array,
similar to https://github.com/asiripanich/emdash/commit/eb7921f1

This change works (see screenshots).
However, it has some hacks that should be cleaned up by somebody with more R knowledge than me.

- Multiple copies of `datatable_options`

The current codebase has two identical copies of `datatable_options`
https://github.com/asiripanich/emdash/blob/5334c9229febaddaa234a3d1adc62f8f90547a90/R/app_server.R#L167
and
https://github.com/asiripanich/emdash/blob/5334c9229febaddaa234a3d1adc62f8f90547a90/R/app_server.R#L195

To be consistent with the existing codebase, I have created a third copy for the trips object.
In the long-term, we need to unify all of these to ensure DRY

- Hardcoded column for `inferred_trips`

The `inferred_trips` column number is currently hard-coded to 5. This is because, due to removed columns before display,

```
inferred_label_index <- which(names(data_geogr$trips) == "inferred_labels")
```
returns 17.

We need to find the index in the list of _displayed_ columns, but I don't
understand R well enough to understand where and how to make that change.

@asiripanich, could you clean up the hacks, or do you want to merge first?